UI 연동작업 완료 - 이제 수정패널&상세보기패널만 구현하면 될듯

+++ process/clearAll.php
@@ -22,7 +22,7 @@ if (!$logged) {
     return;
 }
 
-$success = $this->db()->delete($this->table->todolist)->execute();
+$success = $this->db()->delete($this->table->todolist)->where('admin_idx', '-1')->execute();
 
 $results->success = true;
